added BaseModel to optimize the code of Category, Product models, changed id data type from int to long in DTOs and models, used annotation MappedSuperClass for BaseModel extension into Product and Category models, added GeneratedValue annotaion to make id as autoincrement in BaseModel

+++ src/main/java/com/scaler/productservicespringboot/dto/FakeStorePOSTResponseDTO.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class FakeStorePOSTResponseDTO {
-    private int id;
+    private long id;
     private String title;
     private int price;
     private String description;
@@ -18,7 +18,7 @@ public class FakeStorePOSTResponseDTO {
     public Product toProduct() {
         Product product = new Product();
 
-        product.setId(this.id);
+        product.setId(Long.valueOf(this.id));
         product.setName(this.title);
         product.setDescription(this.description);
         product.setPrice(this.price);
